+++ app/scripts/fhir/sdc-common.js
@@ -97,7 +97,8 @@ function addCommonSDCFns(ns) {
     }
     return ret;
   };
-  
+
+
   /**
    * Do a shallow copy of specified fields from source to target.
    *