+++ CHANGELOG.md
@@ -2,9 +2,15 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
-## [15.6.3] 2019-06-06
+## [15.7.1] 2019-06-06
 ### Fixed
 - Observation-based prepoluation of list fields now works.
+
+## [15.7.0] 2019-06-03
+### Fixed
+- Used a correct extension for coding instructions in FHIR Questionnaire
+### Added
+- Added support for coding instructions format in FHIR Questionnaire
 
 ## [15.6.2] 2019-05-31
 ### Fixed